extracted most of the mainMenu code into its own class then referenced it in main. This will help readability and should make testing easier
