+++ src/mainMenu.h
@@ -1,0 +1,6 @@
+ int launchCraft;
+void initializeMainMenu(int argc, char** argv);
+void renderMainMenu();
+void processNormalKeys(unsigned char key, int x, int y);
+void handleMouseClick(int button, int state,int x, int y);
+void renderButtonBackground();
